Disable SSL certificate verification
This tests' purpose is not to test the certificates in use. Therefore we
can disable the certificate verification so that tests are easier
(possible? See issue #110) to run locally with self-signed certificates.

The comment
https://github.com/kubernetes/k8s.io/pull/19#issuecomment-265556199
suggests that when these tests were first written, certifiacte
verification was disabled by default. So we're confident that the intent
of the tests doesn't rely on working certificate verification.

Closes: #110