dns: remove devstats CNAME entries
The official preferred url is https://k8s.devstats.cncf.io and after
changes recently using the Kubernetes org CNAMEs results in a
certificate error.  No docs appear to reference the vanity url, and since
it now errors and is not a supported config on the cncf/devstats side,
the CNAMEs should be removed.

Signed-off-by: Tim Pepper <tpepper@vmware.com>