Update terraform for GKE clusters
I went through the gcloud docs and mapped every flag to terraform, and
made sure this represents my best understanding of them.

I ran this against the test project.  I think we are just about ready to
do it for real.

Two things I wanted to resolve:

1) Where should we store the state?  It doesn't seem right to ask
everyone to sign up for an account on another service, when we could
store state on GCS.  How?

2) I'd like us to agree on how we want to do multiple clusters in a
single project, because I assume that will be pretty important.