Enable GCB builds for staging repos
Enable GCB API.

Create a scratch bucket for logs and stuff.

Allow the prow svcacct to trigger builds and log.

Remove retention policy for staging buckets (not needed for most and
disallowed for GCB scratch)

After this change, a small prow PR and a YAML file in your repo and you
can pos-submit build and push to staging GCR without human hands
touching it.