add script for prod backups
This is a rough cut of some tooling for basic backups from prod to a
backup. The `backup.sh` script will run from a periodic Prow job every
hour. This script expects the "k8s-artifacts-prod-backup" GCR to exist,
as well as a service account that has write access to this backup
location.