groups: validate length of emails used for staging repos
Ref: https://github.com/kubernetes/k8s.io/pull/289#issuecomment-508806417

For a staging repo, PROJECT_ID is equal to `k8s-staging-<project-name>`.
gcloud only allows PROJECT_ID values between lengths of 6 and 30.

After discounting the number of characaters in `k8s-staging-` (12), we
are left with 18 characters for the project name. This project name is
also used in the email ID for the google group used to provide access to
the staging repo.

To ensure that the project name stays within the character limit, this
commit adds a test to validate that the project name in the email ID
doesn't have more than 18 characters.