infra/gcp: Grant Service Usage Consumer role for GCB users
Users granted access to GCB will not have access to submit GCB jobs
without the serviceusage.services.use permission.

The following error will be displayed:
Caller does not have required permission to use project project:<project>.
Grant the caller the Owner or Editor role, or a custom role with
the serviceusage.services.use permission.

All actions that include a billing project in the request require
serviceusage.services.use permission for the project that's specified.

The Service Usage Consumer role appears to be the least-privilege
default role that grants this permission.

Ref: https://cloud.google.com/storage/docs/access-control/iam-console

Signed-off-by: Stephen Augustus <saugustus@vmware.com>