add backup test
This moves most of the backup logic into a new backup_lib.sh. This logic
is then extended for testing purposes in backup_test.sh (and also tested
there). The backup_prod.sh is the same as the old backup.sh, and is
symlinked from backup.sh so as to not break old jobs.

Please see the README.md for an explanation of how the backup_test.sh
script works.