Fix DNS lame delegation of canary.canary
To fix this, we need to not delegate canary.canary, which means loading
SLIGHTLY different yaml in canary than prod.

This change makes the push script "pre-cook" config files by joining
them together.  These are saved in a temporary dir and octodns reads
from there.  This works because records are top-level items in the YAML
schema.

I could not find a way to make octodns do this itself.

Dry-run shows 2 deletions, as expected.