Revert "Merge pull request #623 from listx/master"
This reverts commit ecfc5398bdb87b57730c807d35e2a804c3bfb56b, reversing
changes made to 670fb0c6ad3000d4c7e5a51d96ebaa4cb8632da6.

This reverts #623. For some reason, with this change there is now an
index out-of-bounds failure in the ci-k8sio-cip job here:
https://prow.k8s.io/view/gcs/kubernetes-jenkins/logs/ci-k8sio-cip/1235320791898263552.

I am puzzled why this did not manifest itself in the post-k8sio-cip
job (postsubmit) when #623 was merged. However, the simple fix is to
revert while I figure this out.