Allow RBAC group members to view members
This is required for RBAC to work.  It used to be true by default, but
it was hanged late last year.  Since then, the people using the cluster
have mostly have cluster-admin access.  I reproduced it by adding my own
gmail to the group and then Mike Danese pointed out this setting was
missing.