releng: Backfill most recent debian-base, debian-iptables, and pause
'pull-kubernetes-cross' expects debian-base and debian-iptables to be in
the same registry. As we haven't promoted a new debian-iptables image
yet, we need to backfill the most recent version into community infra.

We do the same for the debian-base and pause images to support anyone
who wants to move to the new image registry without changing image
versions as well.

Versions backfilled:
- debian-base (all arches): v2.0.0
- debian-iptables (all arches): v12.0.1
- pause (all arches): 3.2

Signed-off-by: Stephen Augustus <saugustus@vmware.com>