change boskos pool name to k8s-infra-gce-project
kubetest defaults to asking for a project of type `gce-project` if it's
not explicitly specified via --gcp-project-type=. So initially I
thought it would be a good idea to use that for the bsokos pool in this
build cluster.

The problem is when we try to use monitoring.prow.k8s.io's boskos
resource dashboard to show this boskos' resources.  There's nothing to
differentiate from k8s-prow-builds' gce-project pool, and this boskos'
gce-project pool.

The easiest thing to do for now is to use uniquely named pools across
boskos instances. This will require that jobs explicitly add a flag
--gcp-project-type=k8s-infra-gce-project when migrating to this
build cluster