kubernetes-csi: promote first multi-arch mock-driver release
csi-test v3.1.1 produced the first multi-arch image. We probably want
to promote only the combined image (tag: v3.1.1) and not the
individual ones for each architecture (tags: amd64-linux-v3.1.1,
ppc64le-linux-v3.1.1, s390x-linux-v3.1.1).

This showed that the `tags~` regex matching does a sub-string match
and not a full match, so the script gets updated a bit to simplify
some terms and tighten the one that matches versions.