Fix prow-build nodepool pool1 initial_count
I scaled this nodepool through the google cloud console when it became
apparent that autoscaling wasn't going to scale up to initial_count
until there was cause for an autoscaling event. It appears doing so
set the nodepool's initial_count, so `terraform plan` showed terraform
wanting to destroy and recreate the nodepool to get back down to an
initial_count of 2.

Updating terraform resource definition to follow so terraform plan
shows no diff