Complete config for scl-image-builder gcr staging
Entries for "scl-image-builder" were hadded to groups.yaml and
ensure-staging-storage.sh 5+ months ago, but the configuration was never
completed and the Google Group was never created. This PR completes the
config according to the latest documentation. Afterwards, the Google
Group can be created and the gcr.io repo can start to be used.