Added AllowWebPosting option to groups reconciler
As currently if you want the group to be accessible not only via gmail
vut also via web interface you would have to set the option "Group
Settings -> Posting Policies -> Conversation: ON" manually, this change
is adding option `AllowWebPosting` to the groups configuration, so
reconciler will set it automatically when your group will have setting
`AllowWebPosting: "true"`.

I also added test to check if groups which should have this option set
(at this point as via #1171 I only added `leads@kubernetes.io` to the
list of groups to validate), and if not, or the value of this option is
not equal `true` it will fail.

As there is already a lot of groups and I didn't want to imperatively
change this setting by default to `false` for them I have not introduced
the default value for that option, but we can change it if we'll decide
that should be the case.

Signed-off-by: Bart Smykla <bsmykla@vmware.com>