infa/gcp/lib: add empower_gke_for_serviceaccount
As an alternative to empower_ksa_to_svcacct that reduces boilerplate on
the caller.

These functions may better belong in lib_iam, and perhaps a better name
is empower_serviceaccount_for_gke, but this is good enough for now. Can
revisit whenever we feel like replacing uses of empower_ksa_to_svcacct