infra/gcp/gcs: use gsutil iam set instead of ch
`gsutil ch` specifically doesn't support convenience bindings (e.g.
a role binding with "projectEditors:project-foo" as a member that will
expand to all members with "roles/editor" on project-foo) as it goes against
the principle of least privilege

Annoyingly, this is the case not just for new bindings, but also
attempts to remove existing bindings. Instead, we must follow a
read-modify-write pattern with `gsutil iam get` and `gsutil iam set`

This was really only needed for ensure_removed_gcs_binding but for
parity I redid ensure_gcs_binding to follow the same pattern. If this
turns out to be wildly slower than before, we should revert for the
common case of adding new bindings.