clusters/aaa: move some iam bindings to infra/gcp
`terraform apply` and `./infra/gcp/ensure-main-project.sh` were getting
in a fight over authoritative IAM bindings for the kubernetes-public
project

IMO terraform shouldn't be setting authoritative bindings for a project
unless it's also managing the project.  Whereas in this case, the module
in question is just for managing a GKE cluster within the project. The
infra/gcp/ensure-main-project.sh script is the source of truth about the
configuration of the kubernetes-public project.

I opted to move the bindings over to ensure-main-project.sh instead of
updating the terraform module to non-authoritatively add IAM members,
since they were related to a group that is managed outside of terraform
vs. service accounts that were created by terraform.

As part of this commit I preemptively ran `terraform state rm` for the
two resources that were removed, so that terraform won't try to destroy
them