OWNERS: Update Release Engineering aliases
release-engineer-approvers/-reviewers:
- Emeritus Jorge, Tim, Pengfei, and Yang

cip-approvers/-reviewers:
- Remove Justin SB
- Add Dan and Sascha

Remove build-image-approvers/-reviewers:
This group was initially created to provider additional approval
coverage for build/base image promotions as we transitioned the image
builds to the Release Engineering subproject.

Now that we have Release Managers actively involved in image building,
we simplify approvals by placing it in the hands of the RelMgr group.

Signed-off-by: Stephen Augustus <foo@auggie.dev>