hack: add k8s-infra-diff-buckets.sh
Diff the contents of two GCS bucket paths via `gsutil ls`. This is
motivated by trying to monitor the delta between google-owned and
community-owned CI artifacts.

The original plan was to rewrite this to use `gsutil rsync --dry-run`
and then remove the `--dry-run` to actually sync the buckets
periodically via a prow job. It turns out the files that motivated this
are not bit-for-bit identical, nor do they have accurate enough
timestamps to be used for comparison, since they are produced by
different jobs running on different clusters at slightly different
times.

So for now, since I've been reporting results from this publicly, get
the script off of my laptop.

Sample usage:

    # diff a single-level listing, discard intermediate files
    ./hack/k8s-infra-diff-buckets.sh \
        gs://kubernetes-release-dev/ci/ \
        gs://k8s-release-dev/ci/

    # diff a recursive listing, leave intermediate files behind
    DEBUG=true ./hack/k8s-infra-diff-buckets.sh \
        gs://kubernetes-release-dev/ci/** \
        gs://k8s-release-dev/ci/**