infra/gcp: rename ensure_api to ensure_services
Since the gcloud command is `gcloud services` let's drop the API
nomenclature, and pluralize to reflect that N services can be enabled at
once now.

Add a corresponding ensure_disabled_services function with a scary
warning.

While following the the rename, I also refactored to use lists of
services if relevant, for:

- ensure-prod-storage.sh
- ensure-prod-storage-gclb.sh
- ensure-env-cip-auditor.sh