infra/gcp: organize constants and includes in lib
Describe the reason includes are sourced in the way they are

Pull more constants into lib.sh and organize them by "purpose" and then
alphabetically within. The following were added to lib, or scripts were
modified to use these:

- RELEASE_ADMINS, RELEASE_WRITERS, RELEASE_VIEWERS
- PROW_GOOGLE_TRUSTED_SERVICE_ACCOUNT (formerly PROW_SVCACCT)
- PROW_BUILD_SERVICE_ACCOUNT
- PROW_TRUSTED_BUILD_CLUSTER_PROJECTS
- PROW_UNTRUSTED_BUILD_CLUSTER_PROJECTS

Shellcheck complains that some of these aren't used within lib.sh and
should be exported if they're used elsewhere. Going to intentially hold
off on that for now, it may turn out not all of these need to be
"global", or that there's a better place to store these.