audit: strip etags from all json
Do this by add a format_gcloud_json function, and passing all gcloud (or
bq, gsutil) output through it.

Format calls so the argument to use json output and the pass through the
function are on the same line to make it clear this should be used
everywhere json output is used.

NOTE: in reformatting the `bq` calls to make this change, I switched
away from prettyjson format, since jq will do the prettifying for us.
However, this means that keys will be stored in a slightly different
order than before.