images: add k8s-infra image
This is intended to be used for running scripts and tests in CI, but may
also serve as a development environment if someone were to mount things
in appropriately.

I'm trying an approach of not copy-pasting the same set of default
values around, so until we need variants, the Dockerfile is the place to
change them.

Setup a `make cloudbuild` target that doesn't push `:latest` for
dev/test purposes

Verified by running `make -C images/k8s-infra build run WHAT=...`

- ./audit/audit-gcp.sh
- ./audit/check-audit-pr.sh
- ./audit/create-or-update-audit-pr.sh
- ./hack/verify.sh
- ./hack/verify-terraform.sh
- ./infra/gcp/ensure-organization.sh
- ./infra/gcp/prow/ensure-e2e-projects.sh

Verified cloudbuild.yaml via `make -C images/k8s-infra cloudbuild`