hack: add verify-executable.sh
Expect that any file that ends in .sh or .py should be executable,
except lib*.sh, which should be sourced. Ignore the same files as git
when searching for files to verify.

This was honestly more bash than I was expecting for this, but I
found I couldn't quite figure out the right `find` incantation to
do everything there... and bash doesn't have a logical xor operator.