Add sbom.k8s.io server entry to nginx configmap
This commit adds two new server configs to the nginx configmap:

sbom.k8s.io/sbom.kubernetes.io

The intent of these URLs is to provide the download locations for the
Kubernetes bill of materials (SBOM). The redirects define two URIs that
map to the release bucket, where the documents reside. Example:

	http://sbom.k8s.io/v1.22.0/release
	http://sbom.k8s.io/v1.22.0/source

These two will redirect to the following locations in the release bucket:

https://storage.googleapis.com/kubernetes-release/release/v1.22.0/kubernetes-release.spdx
https://storage.googleapis.com/kubernetes-release/release/v1.22.0/kubernetes-source.spdx

Signed-off-by: Adolfo García Veytia (Puerco) <adolfo.garcia@uservers.net>