apps/prow: drop DNS record for k8s-infra-prow.k8s.io
Ref: https://github.com/kubernetes/k8s.io/issues/2505

Drop the AAAA record to ensure the certificate for k8s-infra-prow.k8s.io
can be provisioned.
I'm still not sure how managed-certificate-controller is failing to
provide the certificate. This DNS record is not a hard requirement for
this prow instance.
I also keep the IPv6 IP address so we can use once we understood the
issue.

Signed-off-by: Arnaud Meukam <ameukam@gmail.com>