apps/prow: set rollout strategy for ghproxy
Currently any new deployment of ghproxy creates a new pod but fails to
pass the `Running` state.

```
ghproxy-55d56d7cf5-42mf2                  0/1     ContainerCreating   0          158m
ghproxy-74496ccddb-2qt78                  1/1     Running             0          10h
```

This happens because the pod that needs to be terminated is attached to
PV created by the PVC
[ghproxy](https://github.com/kubernetes/k8s.io/blob/main/apps/prow/cluster/ghproxy_deployment.yaml#L16-L30).

Use the `Recreate` rolling strategy should force the pod to be
recreated.

Signed-off-by: Arnaud Meukam <ameukam@gmail.com>