add initial pass at infra diagram
It's likely incomplete in a number of places, some of which I marked
with TODO. This is loosely supposed to represent all of the
infrastructure we manage, and the infrastructure we have yet to migrate.

This was generated from hand written source, but using a google-internal
tool, so I'm not sure if I can share the source from which it was generated

The connections are laid out poorly by the tool, so I wouldn't pay a ton
of attention to them. Roughly speaking solid lines are the path code
flows as it turns into a kubernetes/kubernetes release artifact. The
dotted lines are asynchronous or polling checks. The sample use case I had
in my head was the first cut of an in-development release off of a release
branch, thus including some testgrid dashboard changes.

One thing missing off the top of my head is things that run as Prow Jobs,
so some detail are missing like:
- how k8s-triage is populated
- how k8s-metrics is populated
- how artifact promotion happens
- github automation like label_sync or peribolos

My preference would be to merge and iterate with this, because this is about
as far as I have time to push this for now. I'm open to feedback though,
and interested most specifically in things that are missing from this
diagram, including external services not directly managed by us.