groups: add EmailAddressEquals
Attempt to deal with email addresses that are basically equivalent by
comparing email address with case insensitivity, and stripping dots from
the local part (before the '@').

These rules make sense for gmail addresses, which comprise the majority
of email addresses used in our groups. It's difficult to know whether
they apply universally to all e-mail addresses.

https://www.rfc-editor.org/rfc/rfc5321#section-2.3.11 says that the
semantics for the local part of an e-mail address are determined by the
host. So the only way to know for sure if two email addresses are
equivalent is to send e-mail to both of them, and inspect the resulting
mailbox.

It sure seems like there must be some part of the Directory API that
knows how to handle this equivalence, so I wanted to rely on that
instead of trying to build our own hackish local best guess at email
equivalence.  But when I tried using either the GetMember or HasMember
API calls with the semantically equivalent email addresses, I got back
a 404 not found, and 400 invalid member key respectively.

So, we'll go with this for now until we receive reports of someone not
receiving email due to case-sensitivty or dot-sensitivity issues.