terraform: replace deprecated node_metadata field
When running terraform plan this came across as a deprecation warning
but also as a disagreement with live statement requiring an
update-in-place. It appears this is a perma-diff bug caused by one of
the later versions of terraform-google-provider that we've bumped to.
Upgrading all of the provider versions is a bit more work than just
switching to use the correct non-deprecated field though, so here we
are.