releng: Allow promoter access to canary promotions
Release Engineering will be doing canary promotions of
some projects. These canary releases will not be published
to the publick registries but rather in a different repo
in the staging registry.

To allow these promotions, the promoter needs access to
write to the staging registry.

This PR enables access to the promoter to the first project
that will start issuing canary releases: `k8s-staging-bom`.

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>