Add staging service accounts to `roles/iam.serviceAccountTokenCreator`
The idea is to use the GCB service account to be able to sign images in
staging and promote them later into production. This way we can use
cloudbuild.yaml to specify something like:

```
env:
  GOOGLE_SERVICE_ACCOUNT_NAME=448637284062@cloudbuild.gserviceaccount.com
```

We have to obtain the project ID since it is created randomly for each
staging project.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>