remove non existing image hashes from staging repo
The nfsplugin and smbplugin hashes looks to be unavailable
in staging repo which causes the  generate script to mark for
removal. These hashes would have been promoted to prod but
right now missing in the staging repo. this commit mark those
non existing hashes for removal to avoid disruption of capturing
generate script output.

Signed-off-by: Humble Chirammal <hchiramm@redhat.com>