Temporarily host apt-key.gpg as original url is failing
A cached version of a gpg key is used from the way back machine, dated 2023-02-23 here:

https://web.archive.org/web/20230223152417/https://packages.cloud.google.com/apt/doc/apt-key.gpg

The live gpg key at https://packages.cloud.google.com/apt/doc/apt-key.gpg is inaccessible with a HTTP 500 response.

When the live gpg key is available agian, we can switch the rewrite rule to point to:
https://packages.cloud.google.com/apt/doc/apt-key.gpg

and remove the temporary copy from this repository.

Signed-off-by: Davanum Srinivas <davanum@gmail.com>